+++ lib/kura/extensions.rb
@@ -1,0 +1,3 @@
+# coding: utf-8
+
+require "kura/extensions/google/apis/bigquery_v2/job"